move hmmscan into its own function
